sub_agent/faq_rag_agent/requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client
pydantic
python-dotenv
semantic-kernel>=1.0.0 
semantic-kernel[ollama]
semantic-kernel[mcp]
a2a-sdk[http-server]